+++ packages/grid/src/CellInputField.tsx
@@ -2,7 +2,6 @@ import React, { useCallback, useEffect, useRef, useState } from 'react';
 import classNames from 'classnames';
 import { EMPTY_FUNCTION } from '@deephaven/utils';
 import { SELECTION_DIRECTION } from './GridRange';
-import type { ColumnRestriction } from './GridModel';
 import GridUtils from './GridUtils';
 import './CellInputField.scss';
 
@@ -24,34 +23,6 @@ export type CellInputFieldProps = {
   onContextMenu?: React.MouseEventHandler<HTMLTextAreaElement>;
   style?: React.CSSProperties;
 };
-
-/**
- * A renderer for a single cell input field based on column restriction type.
- *
- * Set `preservesExistingValue = true` on the function to signal that
- * keystroke-initiated edits should open the editor with the existing cell
- * value rather than replacing it with the typed character.
- */
-export type CellInputRendererFn = ((
-  props: CellInputFieldProps & { columnRestrictions: ColumnRestriction[] }
-) => React.ReactNode) & {
-  /**
-   * When true, keystroke-initiated edits preserve the existing cell value
-   * instead of replacing it with the typed character. Intended for renderers
-   * like dropdowns where the typed character has no meaning as a new value.
-   */
-  preservesExistingValue?: boolean;
-};
-
-/**
- * A map from column restriction type string to a cell input renderer function.
- * Grid looks up columnRestrictions[0].type in this registry and falls back to
- * CellInputField when there is no match.
- */
-export type CellInputRendererRegistry = ReadonlyMap<
-  string,
-  CellInputRendererFn
->;
 
 const directionForKey = (key: string): SELECTION_DIRECTION | undefined => {
   switch (key) {
