+++ packages/grid/src/index.ts
@@ -1,5 +1,6 @@
 export { default as CellInputField } from './CellInputField';
 export * from './CellInputField';
+export * from './GridRendererTypes';
 export * from './ColumnHeaderGroup';
 export * from './EditableGridModel';
 export * from './DeletableGridModel';
